api/requirements.txt
gradio
cohere
numpy
sentence-transformers
transformers
python-dotenv
pandas